requirements.txt
absl-py==1.4.0
array-record==0.4.1
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
coverage==7.5.3
cycler==0.12.1
Cython==3.0.10
dataclasses==0.6
decorator==5.1.1
Django==4.2.7
dm-tree==0.1.8
etils==1.5.2
fire==0.6.0
flatbuffers==1.12
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gcloud==0.18.3
gin-config==0.5.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.11.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
joblib==1.4.0
jws==0.1.3
kaggle==1.6.12
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
librosa==0.8.1
llvmlite==0.36.0
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.4.3
neural-structured-learning==1.4.0
numba==0.53.0
numpy==1.23.0
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==20.9
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
promise==2.3
proto-plus==1.23.0
protobuf==3.19.0
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-dateutil==2.9.0.post0
python-jwt==2.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.29.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
resampy==0.4.3
rsa==4.9
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
sqlparse==0.4.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.22.0
tensorflow-datasets==4.9.3
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.1
tensorflow-model-optimization==0.8.0
tensorflowjs==3.18.0
termcolor==2.4.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.3.0
tf-slim==1.1.0
tflite-model-maker==0.3.4
tflite-support==0.4.0
threadpoolctl==3.4.0
toml==0.10.2
tqdm==4.66.2
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.25.11
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
